Improve status management and K8S watcher reliability
- Empty sync_status doesn't exist anymore, sync_pending will be
  the default;
- Observer thread in K8S will now also build the hashtree. Errors
  in building the hashtree will result in status reset instead of
  information loss;
- Move TreeBuilder in the tree_manager module, since it's not DB
  specific;
