+++ aim/tree_manager.py
@@ -13,6 +13,7 @@
 #    License for the specific language governing permissions and limitations
 #    under the License.
 
+import copy
 import traceback
 
 from oslo_log import log as logging
@@ -413,3 +414,119 @@ class TenantHashTreeManager(TenantTreeManager):
             structured_tree.StructuredHashTree,
             AimHashTreeMaker.tenant_rn_funct,
             AimHashTreeMaker.tenant_key_funct)
+
+
+class HashTreeBuilder(object):
+    CONFIG = 'config'
+    OPER = 'oper'
+    MONITOR = 'monitor'
+
+    def __init__(self, aim_manager):
+        self.aim_manager = aim_manager
+        self.tt_maker = AimHashTreeMaker()
+
+    def build(self, added, updated, deleted, tree_map, aim_ctx=None):
+        """Build hash tree
+
+        :param updated: list of AIM objects
+        :param deleted: list of AIM objects
+        :param tree_map: map of trees by type and tenant
+        eg: {'config': {'tn1': <tenant hashtree>}}
+        :return: tree updates
+        """
+
+        # Segregate updates by tenant
+        updates_by_tenant = {}
+        all_updates = [added, updated, deleted]
+        conf = CONFIG_TREE
+        monitor = MONITORED_TREE
+        oper = OPERATIONAL_TREE
+        for idx in range(len(all_updates)):
+            tree_index = 0 if idx < 2 else 1
+            for res in all_updates[idx]:
+                if isinstance(res, aim_status.AciStatus) and aim_ctx:
+                    parent = self.aim_manager.get_by_id(
+                        aim_ctx, res.parent_class, res.resource_id)
+                    # Remove main object from config tree if in sync error
+                    # during an update
+                    if tree_index == 0:
+                        if res.sync_status == res.SYNC_FAILED:
+                            parent = self.aim_manager.get_by_id(
+                                aim_ctx, res.parent_class, res.resource_id)
+                            # Put the object in error state
+                            parent._error = True
+                            all_updates[1].append(parent)
+                        elif res.sync_status == res.SYNC_PENDING:
+                            # A sync pending monitored object is in a limbo
+                            # state, potentially switching from Owned to
+                            # Monitored, and therefore should be removed from
+                            # all the trees
+                            if parent.monitored:
+                                all_updates[-1].append(parent)
+                            else:
+                                all_updates[1].append(parent)
+                        elif res.sync_status == res.SYNCED:
+                            all_updates[1].append(parent)
+                    else:
+                        if parent:
+                            # Delete parent on operational tree
+                            parent_key = self.tt_maker.get_tenant_key(parent)
+                            updates_by_tenant.setdefault(
+                                parent_key, {conf: ([], []), monitor: ([], []),
+                                             oper: ([], [])})
+                            updates_by_tenant[
+                                parent_key][oper][tree_index].append(parent)
+                key = self.tt_maker.get_tenant_key(res)
+                if not key:
+                    continue
+                updates_by_tenant.setdefault(
+                    key, {conf: ([], []), monitor: ([], []), oper: ([], [])})
+                if isinstance(res, aim_status.AciFault):
+                    # Operational Tree
+                    updates_by_tenant[key][oper][tree_index].append(res)
+                else:
+                    if getattr(res, 'monitored', None):
+                        # Monitored Tree
+                        res_copy = copy.deepcopy(res)
+                        updates_by_tenant[key][monitor][tree_index].append(
+                            res_copy)
+                        # Don't modify the original resource in a visible
+                        # way
+                        res = copy.deepcopy(res)
+                        # Fake this as pre-existing
+                        res.pre_existing = True
+                        res.monitored = False
+                    # Configuration Tree
+                    updates_by_tenant[key][conf][tree_index].append(res)
+
+        upd_trees, udp_op_trees, udp_mon_trees = [], [], []
+        for tenant, upd in updates_by_tenant.iteritems():
+            ttree = tree_map[self.CONFIG][tenant]
+            ttree_operational = tree_map[self.OPER][tenant]
+            ttree_monitor = tree_map[self.MONITOR][tenant]
+            # Update Configuration Tree
+            self.tt_maker.update(ttree, upd[conf][0])
+            self.tt_maker.delete(ttree, upd[conf][1])
+            # Clear new monitored objects
+            self.tt_maker.clear(ttree, upd[monitor][0])
+
+            # Update Operational Tree
+            self.tt_maker.update(ttree_operational, upd[oper][0])
+            self.tt_maker.delete(ttree_operational, upd[oper][1])
+            # Delete operational resources as well
+            self.tt_maker.delete(ttree_operational, upd[conf][1])
+            self.tt_maker.delete(ttree_operational, upd[monitor][1])
+
+            # Update Monitored Tree
+            self.tt_maker.update(ttree_monitor, upd[monitor][0])
+            self.tt_maker.delete(ttree_monitor, upd[monitor][1])
+            # Clear new owned objects
+            self.tt_maker.clear(ttree_monitor, upd[conf][0])
+
+            if ttree.root_key:
+                upd_trees.append(ttree)
+            if ttree_operational.root_key:
+                udp_op_trees.append(ttree_operational)
+            if ttree_monitor.root_key:
+                udp_mon_trees.append(ttree_monitor)
+        return upd_trees, udp_op_trees, udp_mon_trees
